app/requirements.txt
fastapi[standard]==0.114.0
uvicorn==0.22.0
pymysql==1.1.1
pydantic==2.0.3
pymongo
python-dotenv
httpx
pytest